refactor(ios): extract composer bar and slash palette from ComposerWorkflowSheet
Behavior-preserving move (ComposerWorkflowSheet.swift 504 -> 141
lines): ChatComposerBar and the slash-command palette
(ComposerSlashItem, ComposerSlashPalette, private palette
view/row/metrics/button-style) move verbatim to
Features/Chat/ChatComposerBar.swift — they are the composer input
surface, used by MobileChatView and the chat placeholder, not sheet
code. The workflow sheet keeps ComposerWorkflowAction, the sheet, its
tile, and ComposerWorkflowSheetHeight. Dropped now-unused UIKit import.
pbxproj wired with explicit entries.
